chore: refactored the code and moved components to core and common dirs
